configure app to use Postgres

+++ src/main/java/com/vlad/restapp/repository/BookRepository.java
@@ -7,9 +7,9 @@ import java.util.List;
 public interface BookRepository {
 
     void add(Book book);
-    Book get(int id);
+    Book get(long id);
     void update(Book book);
-    void delete(int id);
+    void delete(long id);
     List<Book> getAll();
 
 }
